Update to 64bit event number

+++ test/EventStore.ClientAPI.NetCore.Tests/EventNumber.cs
@@ -2,7 +2,7 @@
 {
     public static class EventNumber
     {
-        public const int DeletedStream = int.MaxValue;
-        public const int Invalid = int.MinValue;
+        public const long DeletedStream = long.MaxValue;
+        public const long Invalid = long.MinValue;
     }
 }
